+++ tournament/src/presentation/api/tournament/routes/CreateTournamentRoute.ts
@@ -4,8 +4,8 @@ import { Static, Type } from "@sinclair/typebox";
 import { TournamentApplicationService } from "../../../../application/TournamentApplicationServiceFacade";
 import { PrismaRepositoryFactory } from "../../../../infrastructure/Prisma/PrismaReopsitoryFactory";
 import { PrismaClientProvider } from "../../../../infrastructure/Prisma/PrismaClientProvider";
-import { ToStatic } from "../../../../types/ToStatic";
-import { DIContainer } from "../../../../DIContainer";
+import { ToStatic } from "../../../types/ToStatic";
+import { DIContainer } from "../../../classes/DIContainer";
 import { TournamentDTO } from "../../../../application/dto/TournamentDTO";
 import { TournamentDTO2JSON } from "../TournamentDTO2JSON";
