+++ tournament/src/presentation/interfaces/IDIMachine.ts
