+++ tournament/src/presentation/types/ToStatic.ts
